requirements-dev.txt
pyqt==5.15.7
pytest
flake8
